refactor: use variants for Wordmark component and fix Text collision
Refactors the "boomtick.blog" Wordmark into a centralized component with
predefined variants ('default', 'hero', 'navigation', 'mobile') to
prevent style leakage.

Additionally, resolves a runtime 'TypeError: Failed to construct Text'
by switching to direct, aliased imports for layout primitives to avoid
collisions with the native browser Text constructor.
